+++ Modules/Microsoft365DSC/Examples/Resources/AADAuthenticationMethodPolicyTemporary/1-Create.ps1
@@ -30,7 +30,7 @@ Configuration Example
             Id                       = "TemporaryAccessPass";
             IncludeTargets           = @(
                 MSFT_AADAuthenticationMethodPolicyTemporaryIncludeTarget{
-                    Id = 'DSCGroup'
+                    Id = 'Executives'
                     TargetType = 'group'
                 }
             );